Bug fixes
 - Added sstable verification before WAL is deleted
 - Create new WAL before closing old one, fixing possible flushing error leaving DB unusable
 - Using timestamps to make sure sstable loads in the correct order

+++ flush.go
@@ -24,8 +24,16 @@ func (db *StrataGo) Flush() error {
 	db.immutableMemtable = db.activeMemtable
 	db.activeMemtable = memtable.NewSkipList()
 
+	newWal, err := wal.NewWAL(filepath.Join(db.dataDir, "wal.log"))
+	if err != nil {
+		db.isFlushing = false
+		db.mu.Unlock()
+		return err
+	}
+
 	oldWAL := db.wal
 	if err := oldWAL.Close(); err != nil {
+		newWal.Close()
 		db.isFlushing = false
 		db.mu.Unlock()
 		return err
@@ -34,12 +42,6 @@ func (db *StrataGo) Flush() error {
 	flushingWALPath := filepath.Join(db.dataDir, "wal.log.flushing")
 	os.Rename(filepath.Join(db.dataDir, "wal.log"), flushingWALPath)
 
-	newWal, err := wal.NewWAL(filepath.Join(db.dataDir, "wal.log"))
-	if err != nil {
-		db.isFlushing = false
-		db.mu.Unlock()
-		return err
-	}
 	db.wal = newWal
 	db.mu.Unlock()
 
@@ -65,6 +67,15 @@ func (db *StrataGo) Flush() error {
 	db.immutableMemtable = nil
 	db.isFlushing = false
 	db.mu.Unlock()
+
+	// Verifying SSTable before deleting WAL
+	verifyData, err := reader.ReadAll()
+	if err != nil {
+		return db.handleFlushError(fmt.Errorf("SSTable verification failed: %w", err))
+	}
+	if len(verifyData) != db.immutableMemtable.Size {
+		return db.handleFlushError(fmt.Errorf("SSTable size mismatch"))
+	}
 
 	os.Remove(flushingWALPath)
 	return nil
